+++ strainer/__init__.py
@@ -4,7 +4,9 @@ __author__ = 'Alex Kessinger'
 __license__ = 'Apache 2.0'
 __copyright__ = 'Copyright 2016 Alex Kessiger'
 
-from .structure import (create_serializer, many, child, field)
+from .structure import (create_serializer, many, child, field, dict_field)
 from .exceptions import ValidationException
+from strainer import formatters
+from strainer import validators
 
-__all__ = (create_serializer, many, child, field, ValidationException)
+__all__ = (create_serializer, many, child, field, dict_field, ValidationException, formatters, validators)