+++ tests/test_docs.py
@@ -1,14 +1,14 @@
 from datetime import date
-from strainer import create_serializer, field, child, formatters
+from strainer import serializer, field, child, formatters
 
 
 def test_docs():
 
-    artist_serializer = create_serializer(
+    artist_serializer = serializer(
       field('name')
     )
 
-    album_schema = create_serializer(
+    album_schema = serializer(
       field('title'),
       field('release_date', formatters=[formatters.format_datetime()]),
       child('artist', serializer=artist_serializer)
@@ -31,7 +31,7 @@ def test_docs():
         release_date=date(1971, 12, 17)
     )
 
-    simple_data = album_schema.to_representation(album)
+    simple_data = album_schema.serialize(album)
 
     assert simple_data == {
       'artist': {'name': 'David Bowie'},